SCSIVideo: Add video mirroring device
Receives raw 1bpp screen data as a WRITE(10) command and displays it
on the Raspberry Pi's framebuffer.

This is setup for reading and displaying a compact Mac's 1bpp
512x342 display, but the depth and screen size should probably come
from a SCSI message sent by the remote end during a setup phase.

Some high-resolution logging shows it taking about 83 milliseconds
to receive the 21888 bytes of data through SCSI, and about 3
milliseconds to convert it and draw it on the framebuffer.